mini feat and fix: added new functions for receiving, slightly reworked the independent test

+++ src/quiz/entities/quiz.entity.ts
@@ -229,6 +229,6 @@ export class QuizSummary {
   @Field(() => String, { nullable: true })
   description?: string;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   totalPercents?: number;
 }
